Refactoring ComponentBinding
- move ViewTransition in own file to kept ComponentBindingAdapter clearer
